# openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: ADRO Control API
  version: 0.1.0
  description: Provider-independent Agentic Delivery & Release Orchestrator API.
  license:
    name: Apache License 2.0
    identifier: Apache-2.0
servers:
- url: "/"
security:
- bearerAuth: []
- cookieAuth: []
paths:
  "/api/v1/execution-plans":
    get:
      summary: List immutable execution plans in a workspace
      responses:
        '200':
          description: Execution plans
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/JsonObject"
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/RequestID"
        default:
          description: Problem Details response
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/RequestID"
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
            application/json:
              schema:
                "$ref": "#/components/schemas/JsonObject"
  "/api/v1/execution-plans/{plan_id}":
    get:
      summary: Get a frozen execution plan
      parameters:
      - name: plan_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Execution plan
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/JsonObject"
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/RequestID"
        default:
          description: Problem Details response
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/RequestID"
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
            application/json:
              schema:
                "$ref": "#/components/schemas/JsonObject"
...
  "/api/v1/execution-plans/validate":
    post:
      summary: Validate a graph execution plan without executing it
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/JsonObject"
      responses:
        '200':
          description: Validated graph and deterministic digest
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/JsonObject"
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/RequestID"
            Idempotency-Replayed:
              "$ref": "#/components/headers/IdempotencyReplayed"
        '422':
          description: Graph validation failure
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/RequestID"
        default:
          description: Problem Details response
          content:
            application/problem+json:
              schema:
                "$ref": "#/components/schemas/Problem"
            application/json:
              schema:
                "$ref": "#/components/schemas/JsonObject"
          headers:
            X-Request-ID:
              "$ref": "#/components/headers/RequestID"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: adro_session
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Optional for POST, PUT, and PATCH. Scope is tenant + method + path.
        Reuse with the same query, content type, and body replays the first completed
        response; reuse with different input returns 409 idempotency_key_conflict.
      schema:
        type: string
        minLength: 1
        maxLength: 255
  schemas:
...
    JsonObject:
      type: object
      description: Versioned JSON resource or page. Resource-specific fields remain
        backward-compatible within /api/v1.
      additionalProperties: true
    Problem:
      type: object
      required:
      - type
      - title
      - status
      - detail
      - error_code
      - request_id
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        error_code:
          type: string
        request_id:
          type: string
      additionalProperties: true
...
  headers:
    RequestID:
      description: Stable correlation identifier for this HTTP request.
      schema:
        type: string
    IdempotencyReplayed:
      description: True when the status, headers, and body were replayed from the
        first completed request.
      schema:
        type: string
        const: 'true'